refactor[#22]: agregar __main__ en config_modifier.py y optimizar los tests
